refactor: Add VideoPlayer interaction logic and Change styles
